# docs/swagger.yaml
basePath: /
definitions:
  github_com_fluentfox_api_pkg_response.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  github_com_fluentfox_api_pkg_response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorBody'
    type: object
  internal_admin.AdminUserResponse:
    properties:
      banned:
        type: boolean
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  internal_admin.BanUserRequest:
    properties:
      reason:
        maxLength: 500
        type: string
    required:
    - reason
    type: object
  internal_admin.StatsResponse:
    properties:
      active_today:
        type: integer
      chapters_published:
        type: integer
      total_reviews:
        type: integer
      total_users:
        type: integer
    type: object
  internal_auth.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  internal_auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  internal_auth.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  internal_auth.RegisterRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
      username:
        maxLength: 30
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  internal_auth.RegisterResponse:
    properties:
      message:
        type: string
      user_id:
        type: string
    type: object
  internal_auth.ResetPasswordRequest:
    properties:
      password:
        maxLength: 72
        minLength: 8
        type: string
      token:
        type: string
    required:
    - password
    - token
    type: object
  internal_auth.TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        description: seconds until access token expires
        type: integer
      refresh_token:
        description: raw opaque value sent to client
        type: string
    type: object
  internal_auth.VerifyEmailRequest:
    properties:
      token:
        type: string
    required:
    - token
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    Japanese language learning API — chapters, SRS flashcard reviews, quizzes, XP levelling, streaks, and in-app shop.
    All successful responses are wrapped in {"data": ...}. Paginated responses also include {"meta": {...}}.
    All error responses follow {"error": {"code": "...", "message": "..."}}.
  title: FluentFox API
  version: "1.0"
paths:
  /admin/stats:
    get:
      description: Returns platform-wide statistics (total users, active today, reviews,
        published chapters). Requires admin role.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_admin.StatsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get platform stats
      tags:
      - admin
  /admin/users:
    get:
      description: Returns a paginated list of all users with admin-only fields. Requires
        admin role.
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Results per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_admin.AdminUserResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - admin
  /admin/users/{id}/ban:
    post:
      consumes:
      - application/json
      description: Bans a user account with a mandatory reason. Requires admin role.
      parameters:
      - description: Target user UUID
        in: path
        name: id
        required: true
        type: string
      - description: Ban reason
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_admin.BanUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ban a user
      tags:
      - admin
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Sends a password reset link to the email if an account exists.
        Always returns 200 to prevent email enumeration.
      parameters:
      - description: Account email
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "429":
          description: Rate limited
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      summary: Request password reset
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates credentials and returns access + refresh tokens.
      parameters:
      - description: Login payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "429":
          description: Rate limited
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      summary: Log in
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Revokes the supplied refresh token.
      parameters:
      - description: Refresh token to revoke
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Log out
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchanges a valid refresh token for a new access + refresh token
        pair.
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      summary: Refresh tokens
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a user account and sends an email verification link.
      parameters:
      - description: Registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_auth.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "409":
          description: Email already in use
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "429":
          description: Rate limited
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      summary: Register a new account
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Resets the user's password using a one-time token from the reset
        email.
      parameters:
      - description: Reset token and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Token invalid, expired, or already used
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      summary: Reset password
      tags:
      - auth
  /auth/verify-email:
    post:
      consumes:
      - application/json
      description: Marks the user's email as verified using the token sent during
        registration.
      parameters:
      - description: Verification token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Token invalid, expired, or already used
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      summary: Verify email address
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer <your_access_token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"